requirements.txt
flask>=3.0.0
flask-cors
flask-limiter>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
icalendar
python-dateutil
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
supabase>=2.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
